unordered map with a template for the user options and a password set
added in a vector to take multiple password inputs from the user. also added in the functionality to disable the password being displayed on the terminal unless they funnel it in as an argument. still need to flush out the other options and make a helperclass with various functions. builder script will call the makefile but needs to be flushed out for the user options and to create .deb and .rpm files